// front/src/app/models/tablas-valores.interface.ts
export interface TablasValores {
    id: number | null;
    nombre: string;
    id_modulo: number | null;
    id_usuario: number | null;
    fecha_creacion: Date | null;
    fecha_modificacion: Date | null; 
    id_usuario_modificacion: number|null;
    usuario: string|null;
    usuario_modificacion: string|null;
    }